common commands; job process run for at least 5 seconds before restarting

+++ src/Job.php
@@ -19,6 +19,7 @@ class Job implements ShouldQueue
 
     public function handle() {
         if ( config('lumi-core.enable_job_process_restart') && !isset($GLOBALS['__jobs_count']) ) {
+            $GLOBALS['__jobs_start_time'] = time();
             $GLOBALS['__jobs_count'] = 0;
         }
 
@@ -29,8 +30,14 @@ class Job implements ShouldQueue
             $GLOBALS['__jobs_count']++;
             $max_jobs_count = config('lumi-core.max_jobs_per_process:'.$this->job->getQueue()) ?? config('lumi-core.max_jobs_per_process');
 
-            if ( $GLOBALS['__jobs_count'] == $max_jobs_count ) {
-                \DB::table('jobs')->where('id', $this->job->getJobId())->delete();
+            //check if max jobs count was reached and the process has been running for at least 5 seconds
+            if ( $GLOBALS['__jobs_count'] >= $max_jobs_count && (time() - $GLOBALS['__jobs_start_time']) > 5 ) {
+                $job_id = $this->job->getJobId();
+
+                if ( $job_id ) {
+                    \DB::table('jobs')->where('id', $this->job->getJobId())->delete();
+                }
+
                 die();
             }
         }
